website: route-level code splitting to fix oversized bundle
- lazy-load all route components (React.lazy + Suspense)
- extract defaultSampler2D helpers out of transform.js so the entry
  no longer pulls the GLSL compiler chain (and no longer creates a
  WebGL validation context on every page)
- extract gallery from/to images so Intro doesn't import Gallery
