+++ packages/website/src/Intro.jsx
@@ -7,8 +7,8 @@ import { transitionsOrderByRandom, transitionsByName } from "./data";
 import { githubRepoPath } from "./conf";
 import GlslCode from "./GlslCode";
 import TransitionAuthorAndName from "./TransitionAuthorAndName";
-import { defaultSampler2D } from "./transform";
-import { fromImage, toImage } from "./Gallery";
+import { defaultSampler2D } from "./defaultSampler2D";
+import { fromImage, toImage } from "./galleryImages";
 import { FaGithub } from "react-icons/fa";
 import "./Intro.css";
 import cut1mp4 from "./videos/sintel/cut1.mp4";
