+++ packages/website/src/galleryImages.js
@@ -1,0 +1,6 @@
+// shared by Gallery and Intro; kept apart so Intro doesn't import the Gallery chunk
+import fromImageUrl from "./images/600x400/barley.jpg";
+import toImageUrl from "./images/600x400/hBd6EPoQT2C8VQYv65ys_White_Sands.jpg";
+
+export const fromImage = fromImageUrl;
+export const toImage = toImageUrl;
