+++ packages/website/src/transform.js
@@ -1,30 +1,5 @@
 import { transformSource, createWebGLCompiler } from "gl-transition-utils";
-import defaultSampler2DUrl from "./textures/luma/spiral-2.png";
-
-export const defaultSampler2D = defaultSampler2DUrl;
-
-export function defaultSampler2DParamsForType(types) {
-  let res;
-  Object.keys(types).forEach(key => {
-    if (types[key] === "sampler2D") {
-      if (!res) res = {};
-      res[key] = defaultSampler2D;
-    }
-  });
-  return res;
-}
-
-export function supplyDefaultSampler2DToTransition(transition) {
-  const sampler2DParams = defaultSampler2DParamsForType(transition.paramsTypes);
-  if (!sampler2DParams) return transition;
-  return {
-    ...transition,
-    defaultParams: {
-      ...transition.defaultParams,
-      ...sampler2DParams,
-    },
-  };
-}
+import { supplyDefaultSampler2DToTransition } from "./defaultSampler2D";
 
 const canvas = document.createElement("canvas");
 // the size need to be > 256 just so we can be accurate enough on colors
